Reland "Origin policy: implement window.originPolicyIds"
This plumbs the origin policy IDs from the net-side OriginPolicyContents
through to the renderer, where they get exposed on Window. This does not
yet tackle WorkerGlobalScope, but it does add idlharness tests for it,
which fail for now.

This is a re-land of
https://chromium-review.googlesource.com/c/chromium/src/+/2089994 which
was reverted in
https://chromium-review.googlesource.com/c/chromium/src/+/2093453. The
re-land adds MSAN test expectations per https://crbug.com/856601.

Bug: 1057123
Change-Id: I30053986f3dfc634399c8e0f3fc1578e062c67f0